[#3] Move files
