+++ Backend/src/Model/Request/Movie/MovieSearchRequestDto.php
@@ -18,38 +18,29 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 declare(strict_types=1);
 
-namespace App\Model\Request\Search;
+namespace App\Model\Request\Movie;
 
 use Symfony\Component\Serializer\Attribute\SerializedName;
 use Symfony\Component\Validator\Constraints as Assert;
 
-readonly class TvSearchRequestDto
+readonly class MovieSearchRequestDto
 {
     public function __construct(
         #[Assert\NotBlank(
             message: 'Search query parameter "q" is required.'
         )]
         public ?string $q = null,
-        #[Assert\Range(
-            notInRangeMessage: 'The year must be between 1000 and 9999.',
-            min: 1000,
-            max: 9999
-        )]
-        #[SerializedName('first_air_date_year')]
-        public ?int $firstAirDateYear = null,
         #[SerializedName('include_adult')]
         public bool $includeAdult = true,
         public string $language = 'de-DE',
+        #[SerializedName('primary_release_year')]
+        public ?string $primaryReleaseYear = null,
         #[Assert\Range(
             minMessage: 'Page number must be at least 1.',
             min: 1
         )]
         public int $page = 1,
-        #[Assert\Range(
-            notInRangeMessage: 'The year must be between 1000 and 9999.',
-            min: 1000,
-            max: 9999
-        )]
+        public ?string $region = null,
         public ?int $year = null
     ){}
 }
